Refactored to shift previous code under src into a new package under src.
