+++ pyproject.toml
@@ -19,3 +19,8 @@ authors = [
 description = "Python implementation of Referent Tracking 2.0"
 readme = "README.md"
 requires-python = ">=3.11"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "src/rt_core_v2"
+]
