+++ src/rt_core_v2/metadata_accessory.py
