+++ tests/test_ntoxtuple.py
@@ -1,11 +1,11 @@
-from src.ids_codes import Rui
-from src.rttuple import Atuple, NtoR, NtoN, NtoDE, NtoC, NtoLackR
+from ids_codes import Rui
+from rttuple import Atuple, NtoR, NtoN, RuiStatus, NtoDE, NtoC, NtoLackR
 
 def print_atuple(a):
 	print("A tuple information:")
 	print("\truit: the rui that denotes tuple itself", str(a.ruit.uuid))
 	print("\truip: rui that was assigned to some PoR", str(a.ruip.uuid))
-	print("\tis ruip reserved? ", a.ruip.is_reserved())
+	print("\tis ruip reserved? ", a.is_reserved())
 	print('\tis ruip singularly unique vs. potentially non-singularly unique: ', a.unique)
 	print("\truia: rui that denotes person who assigned ruip to some PoR", str(a.ruia.uuid))
 	print("\tt: time that ruia assigned/reserved ruip to/for some PoR", str(a.t))
@@ -34,32 +34,32 @@ def print_nton_tuple(nton):
 	print()
 
 # Rui that represents the person authoring these tuples
-a = Rui.Rui(Rui.RuiStatus.assigned)
+a = Rui.Rui()
 # Rui that stands for the repeatable PoR called "human being"
-h = Rui.Rui(Rui.RuiStatus.assigned)
+h = Rui.Rui()
 # Rui that stands for interval over which author has been instance of human being
-tr1 = Rui.Rui(Rui.RuiStatus.assigned)
+tr1 = Rui.Rui()
 k = Atuple(tr1, ruia=a)
 
 ntor = NtoR(None, a, True, "instance of", h, tr1)
 print_ntor_tuple(ntor)
 
 #let x be the RUI standing for Kuala Lumpur
-x = Rui.Rui(Rui.RuiStatus.assigned)
+x = Rui.Rui()
 y = Atuple(x)
 print_atuple(y)
 
-q = Rui.Rui(Rui.RuiStatus.reserved)
-z = Atuple(q, ruia=a, unique="+SU", ar='R')
+q = Rui.Rui()
+z = Atuple(q, ruia=a, unique="+SU", ar=RuiStatus.reserved)
 print_atuple(z)
 
 #let s be the RUI standing for the territory of Malaysia
-s = Rui.Rui(Rui.RuiStatus.assigned)
+s = Rui.Rui()
 w = Atuple(s, ruia=a, unique="+SU")
 print_atuple(w)
 
 #let tr2 be interval over which kuala lumpur has been part of Malaysia
-tr2 = Rui.Rui(Rui.RuiStatus.assigned)
+tr2 = Rui.Rui()
 j = Atuple(tr2, ruia=a)
 nton = NtoN(None, x, True, "part of", [x, s], tr2)
 print_nton_tuple(nton)
